feat(drawer): create `Drawer` comp and fix `PageTopBar` bgColor
